Only work with named files, not stdin
While this is removing functionality, it enables me to work with a
philosophical mindset more conducive to what I'm trying to accomplish
with this particular program: to quickly identify modules, instruments
and samples in a list of files.  I can't imagine why anyone would
be piping any of these.  This will also let me eventually weed out
false positive matches (not that I've noticed any myself, but it's
bad design philosophy to allow them at all).